requirements.txt
flask>=2.3
flask-sqlalchemy>=3.1
apscheduler>=3.10
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
gunicorn>=21.0
